Move bind operator to misc package.
